+++ src/Lego/Field/Concerns/HasScope.php
@@ -3,18 +3,19 @@
 use Lego\Operator\Query\Query;
 
 /**
- * Class ScopeOperator
+ * Class HasScope
  * @package Lego\Field\Operators
  */
-trait ScopeOperator
+trait HasScope
 {
-    private $scope;
+    protected $scope;
 
     /**
      * 自定义的过滤逻辑
      *
      * $scope 可以为 string 或 Closure
      *  - string ：见 Eloquent Query Scopes
+     *  - null ：同上 string ，自动使用当前 field name 作为 scope 函数名
      *  - Closure ：传入的 Closure 接收两个参数
      *      - $query , instanceof \Lego\Operator\Query\Query ，注意：此处不是 Laravel 中的 Query Builder
      *      - $value , 此项的输入值
@@ -41,11 +42,11 @@ trait ScopeOperator
      * @param string|\Closure $scope
      * @return $this
      */
-    public function scope($scope)
+    public function scope($scope = null)
     {
-        lego_assert(is_string($scope) || $scope instanceof \Closure, 'illegal $scope');
+        lego_assert(is_null($scope) || is_string($scope) || $scope instanceof \Closure, 'illegal $scope');
 
-        $this->scope = $scope;
+        $this->scope = $scope ?: $this->name();
 
         return $this;
     }