+++ src/Lego/Widget/Form.php
@@ -147,6 +147,10 @@ class Form extends Widget implements HasMode
     private function saveFieldsValueToStore()
     {
         $this->editableFields()->each(function (Field $field) {
+            if ($field->isDoesntStore()) {
+                return;
+            }
+
             $field->syncValueToStore();
         });